backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.22
psycopg[binary]==3.2.4
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
openai-agents>=0.0.7
mcp>=1.2.0
openai>=1.60.0
httpx>=0.27.0
